% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_parser.R
\name{parse_swimming}
\alias{parse_swimming}
\title{Parse swimming summary data}
\usage{
parse_swimming(workouts)
}
\arguments{
\item{workout}{A tibble containing workout data}
}
\value{
a tibble containing summarized data of each swimming session
}
\description{
Filter the swimming data form Moov workout data.
    Then, extracts JSON data related to swimming
}
\examples{
\dontrun{
workouts <- read_moov("user.db")
parse_swimming(workouts)
}
}
