% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/session_parser.R
\name{plot_sensor_swim}
\alias{plot_sensor_swim}
\title{Plot swimming sensor data in a given index range}
\usage{
plot_sensor_swim(data, from = NA, to = NA)
}
\arguments{
\item{data}{A tibble containing sensor data @seealso \code{\link{read_moov_swim_bin}}}

\item{from, to}{An integer of the range of the indices to plot. If missing,
plot from the beginning to the end of the data.}
}
