% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_parser.R
\name{plot_swim_summary}
\alias{plot_swim_summary}
\title{Plot swimming stats over session}
\usage{
plot_swim_summary(swim_data, col_name, ylim = c(1.5, 2.4))
}
\arguments{
\item{swim_data}{A tibble containing the swimming summary data  (@seealso \code{\link{parse_swimming}}).}

\item{col_name}{The name of the column to plot. Interesting names are:
\code{"distance_per_stroke", "stroke_rate"}. For all names, use \code{names(swim_data)}.}

\item{ylim}{A vector of two number to limits the y-axis of the graph}
}
\examples{
\dontrun{
plot_swim_summary(swim_data, "distance_per_stroke") + geom_smooth(method="loess")
}
}
