% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_parser.R
\name{read_moov}
\alias{read_moov}
\title{Read Moov sensor data}
\usage{
read_moov(db_path)
}
\arguments{
\item{db_path}{The path of the database file \code{user.db}}
}
\value{
a tibble containing data of each workout session
}
\details{
To use this function, you need to find the \code{user.db} fil from the dump of
MOOV data (@seealso \code{\link{read_moov_swim_bin}}) at
\code{/Applications/moov.core/Library/Application Support/user_data/( user id )/}.
}
\examples{
\dontrun{
read_moov("user.db")
}
}
