% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/session_parser.R
\name{read_moov_swim_bin}
\alias{read_moov_swim_bin}
\title{Load the MOOV sensor data from a swimming session}
\usage{
read_moov_swim_bin(id, swimming_root)
}
\arguments{
\item{id}{session ID, e.g., \code{"swimming-k5y5pqbib17vbjqrh9gr16e"}}

\item{swimming_root}{path to the "swimming" folder (see the Details section)}
}
\value{
a two-column tibble. `x`: sequence number, `y`: value from the binary
}
\description{
Find the "swimming_result.pb" for the given session \code{id}. Load and
extract the part that contains sensor data.
}
\details{
To use this function, you need to extract the "swimming" folder from your
iOS backup. One solution is to use iExplorer (\url{https://macroplant.com/iexplorer}).
(The free version allows you to extract unlimited number of files.)
The \code{swimming_root} folder is in:
\code{/Applications/moov.core/Library/Application Support/user_data/( user id )/swimming}

I haven't figured out the acutal meaning of the data. Heck, I don't know whether
I extracted it correctly! However, preliminary plots shows that the data are periodic.
The data seems like what we would get from accelerometer in lap swimming.
The periodicity of strokes and laps are present in this data.

Known limitation: this code cannot parse recent swim logs that use different header.
I already figured out the binary that marks the beginning of the log (see the source code).
However, I haven't got around to add that part in code yet.
}
\examples{
\dontrun{
read_moov_swim_bin("swimming-a2tj48knt09_f0pkqin8vnt", "data/swimming")
}
}
